fixed Gender

+++ my-app/src/components/Customers/GenerateId.js
@@ -1,8 +1,13 @@
 let idCounter = parseInt(localStorage.getItem('lastIdCounter')) || 90910001;
 
-export const GenerateId = () => {
-    const newId = idCounter++;
-    localStorage.setItem('lastIdCounter', idCounter.toString() );
-    return newId;
+export const GenerateId = (existingId) => {
+  if (existingId) {
+    return existingId; 
+  }
+    else {
+      const newId = idCounter++;
+      localStorage.setItem('lastIdCounter', idCounter.toString() );
+      return newId;
+    }
   };
